agent/requirements.txt
bedrock-agentcore>=0.1.0
bedrock-agentcore-starter-toolkit>=0.1.21
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.35.0